+++ tests/Command/DebugRouterCommandTest.php
@@ -66,6 +66,52 @@ class DebugRouterCommandTest extends TestCase
         $this->assertSame(1, $commandTester->getStatusCode());
     }
 
+    public function testCommandDescribesANamedRouteForARouterWhenGivenTheRouterNameAsAnArgument(): void
+    {
+        $command = new DebugRouterCommand($this->buildRegistryWithValidRouter());
+
+        $commandTester = new CommandTester($command);
+        $commandTester->execute(
+            [
+                'router' => 'test',
+                'route' => 'user_chat',
+            ]
+        );
+
+        $this->assertStringEqualsFile(__DIR__.'/../Fixtures/command_output/valid_route.txt', $commandTester->getDisplay());
+    }
+
+    public function testCommandDescribesANamedRouteForARouterWhenGivenTheRouterNameAsAnOption(): void
+    {
+        $command = new DebugRouterCommand($this->buildRegistryWithValidRouter());
+
+        $commandTester = new CommandTester($command);
+        $commandTester->execute(
+            [
+                '--router_name' => 'test',
+                'router' => 'user_chat',
+            ]
+        );
+
+        $this->assertStringEqualsFile(__DIR__.'/../Fixtures/command_output/valid_route.txt', $commandTester->getDisplay());
+    }
+
+    public function testCommandRaisesErrorIfNamedRouteDoesNotExist(): void
+    {
+        $command = new DebugRouterCommand($this->buildRegistryWithValidRouter());
+
+        $commandTester = new CommandTester($command);
+        $commandTester->execute(
+            [
+                'router' => 'test',
+                'route' => 'missing',
+            ]
+        );
+
+        $this->assertTrue(false !== strpos($commandTester->getDisplay(), 'The "missing" route does not exist on the "test" router.'));
+        $this->assertSame(1, $commandTester->getStatusCode());
+    }
+
     private function buildRegistryWithValidRouter(): RouterRegistry
     {
         $router = new Router(